# Conflicts:
#	Resources/views/frontend/components/product/product-list-item/layouts/product-list-item-layout-1/infor.blade.php

+++ Http/Livewire/Index/Filters/RangePrices.php
@@ -25,7 +25,7 @@ class RangePrices extends Component
     */
 	public function mount()
 	{
-    	$this->step = setting('icommerce::filterRangePricesStep',null,20000);
+    	$this->step = (int)setting('icommerce::filterRangePricesStep',null,20000);
 		$this->priceMin = 0;
 		$this->priceMax = 1;
 		$this->selPriceMin = 0;
